Revert "httpd: upgrade.c: revert change to uprade.c"
This reverts commit e22739ee492b131600f9b6504e6161832f3518d6.

+++ tomato64/package/libshared/shared/shutils.c
@@ -493,9 +493,9 @@ safe_fread(void *ptr, size_t size, size_t nmemb, FILE *stream)
  * @return	number of items successfully written
  */
 int
-safe_fwrite(const void *ptr, size_t size, size_t nmemb, FILE *stream)
+safe_fwrite(const void *ptr, size_t size, int nmemb, FILE *stream)
 {
-	size_t ret = 0;
+	int ret = 0;
 
 	do {
 		clearerr(stream);
